feat: rename models
