+++ packages/core/src/models/file.ts
@@ -1,69 +1,8 @@
-import { action, computed, makeObservable, observable, toJS } from 'mobx';
-import { getValue, isNil, logger, setValue } from '@music163/tango-helpers';
-import type { FileType, IFileConfig } from '../types';
-import { formatCode } from '../helpers';
+import { action, computed, makeObservable, observable } from 'mobx';
+import { isNil } from '@music163/tango-helpers';
+import type { IFileConfig } from '../types';
 import { AbstractWorkspace } from './abstract-workspace';
-
-/**
- * 普通文件抽象基类，不进行 AST 解析
- */
-export abstract class AbstractFile {
-  readonly workspace: AbstractWorkspace;
-  /**
-   * 文件名
-   */
-  readonly filename: string;
-
-  /**
-   * 文件类型
-   */
-  readonly type: FileType;
-
-  /**
-   * 最近修改的时间戳
-   */
-  lastModified: number;
-
-  /**
-   * 文件解析是否出错
-   */
-  isError: boolean;
-
-  /**
-   * 文件解析错误消息
-   */
-  errorMessage: string;
-
-  _code: string;
-  _cleanCode: string;
-
-  get code() {
-    return this._code;
-  }
-
-  // FIXME: cleanCode 是不是只有 viewFile 有 ????
-  get cleanCode() {
-    return this._cleanCode;
-  }
-
-  constructor(workspace: AbstractWorkspace, props: IFileConfig, isSyncCode = true) {
-    this.workspace = workspace;
-    this.filename = props.filename;
-    this.type = props.type;
-    this.lastModified = Date.now();
-    this.isError = false;
-
-    // 这里主要是为了解决 umi ts 编译错误的问题，@see https://github.com/umijs/umi/issues/7594
-    if (isSyncCode) {
-      this.update(props.code);
-    }
-  }
-
-  /**
-   * 更新文件内容
-   */
-  abstract update(code?: string): void;
-}
+import { AbstractFile } from './abstract-file';
 
 export class TangoFile extends AbstractFile {
   constructor(workspace: AbstractWorkspace, props: IFileConfig) {
@@ -85,125 +24,5 @@ export class TangoFile extends AbstractFile {
       this._cleanCode = code;
     }
     this.workspace.onFilesChange([this.filename]);
-  }
-}
-
-export class TangoJsonFile extends AbstractFile {
-  _object = {};
-
-  /**
-   * @deprecated 使用 file.json 代替
-   */
-  get object() {
-    return toJS(this._object);
-  }
-
-  get json() {
-    return toJS(this._object);
-  }
-
-  constructor(workspace: AbstractWorkspace, props: IFileConfig) {
-    super(workspace, props, false);
-    this.update(props.code);
-    makeObservable(this, {
-      _code: observable,
-      _cleanCode: observable,
-      _object: observable,
-      code: computed,
-      cleanCode: computed,
-      object: computed,
-      json: computed,
-      update: action,
-      setValue: action,
-    });
-  }
-
-  update(code?: string) {
-    this.lastModified = Date.now();
-
-    if (isNil(code)) {
-      // 基于最新的 json 同步代码
-      let newCode = JSON.stringify(this._object);
-      try {
-        newCode = formatCode(newCode, 'json');
-      } catch (err) {
-        logger.error(err);
-        return;
-      }
-      this._code = newCode;
-      this._cleanCode = newCode;
-    } else {
-      try {
-        // 基于传入的代码，同步 json 对象
-        code = formatCode(code, 'json');
-      } catch (err) {
-        logger.error(err);
-        return;
-      }
-      this._code = code;
-      this._cleanCode = code;
-      try {
-        const json = JSON.parse(code);
-        this._object = json;
-      } catch (err) {
-        logger.error(err);
-      }
-    }
-    this.workspace.onFilesChange([this.filename]);
-  }
-
-  /**
-   * 根据路径取值
-   * @param valuePath
-   * @returns
-   */
-  getValue(valuePath: string) {
-    return getValue(this.json, valuePath);
-  }
-
-  /**
-   * 根据路径设置值
-   * @param valuePath
-   * @param visitor
-   */
-  setValue(valuePath: string, visitor: (targetValue: any) => any) {
-    const target = this.getValue(valuePath);
-    let next: unknown;
-    if (typeof visitor === 'function') {
-      next = visitor?.(target);
-    } else {
-      next = visitor;
-    }
-    if (next !== undefined) {
-      setValue(this._object, valuePath, next);
-    }
-    return this;
-  }
-
-  /**
-   * 根据路径删除值
-   * @param valuePath
-   * @param visitor
-   */
-  deleteValue(valuePath: string) {
-    const pathList = valuePath.split('.');
-    const lastPath = pathList.pop();
-    const parentPath = pathList.join('.');
-    let target;
-    if (parentPath) {
-      target = this.getValue(parentPath);
-    } else {
-      target = this.json;
-    }
-    if (!target) {
-      return this;
-    }
-    delete target[lastPath];
-    if (parentPath) {
-      this.setValue(parentPath, target);
-    } else {
-      this._object = target;
-    }
-    return this;
   }
 }
