+++ packages/core/src/models/index.ts
@@ -6,12 +6,12 @@ export * from './select-source';
 export * from './drag-source';
 export * from './history';
 export * from './file';
-export * from './module';
-export * from './entry-module';
-export * from './route-module';
-export * from './service-module';
-export * from './store-module';
-export * from './view-module';
-export * from './component-module';
-export * from './node';
+export * from './js-file';
+export * from './js-app-entry-file';
+export * from './js-route-config-file';
+export * from './js-service-file';
+export * from './js-store-entry-file';
+export * from './js-view-file';
+export * from './js-local-components-entry-file';
+export * from './view-node';
 export * from './interfaces';
