+++ packages/core/src/types.ts
@@ -8,7 +8,7 @@ export type SimulatorMode = 'desktop' | 'tablet' | 'phone';
 export enum FileType {
   // js 文件
   Module = 'module',
-  AppEntryModule = 'appEntryModule',
+  JsAppEntryFile = 'jsAppEntryFile',
   StoreEntryModule = 'storeEntryModule',
   RouteModule = 'routeModule',
   ServiceModule = 'serviceModule',
